+++ src/components/WorldMap.jsx
@@ -1,66 +1,64 @@
 import React from 'react';
 
-import Player from './Player';
-import Stats from './Stats';
-
-const styles = {
-  worldMap: {
-    display: 'block',
-    position: 'relative',
-    height: '100%',
-    width: '100%',
-  },
-  worldMapImage: {
-    height: '100%',
-    width: '100%',
-    zIndex: -1,
-  },
-};
+const MAP_SIZE = 2048;
 
 export class WorldMap extends React.Component {
-  static displayName = 'WorldMap';
+  static displayName = 'CanvasWorldMap';
 
   static propTypes = {
     players: React.PropTypes.array,
+    // mapData: React.PropTypes.any,
   }
 
-  playerHovered(player) {
-    this.setState({
-      hoveredPlayer: player,
+  componentDidMount() {
+    const image = new Image();
+    image.onload = this.loadedImage.bind(this, image);
+    image.src = 'CompanionWorldMap.png';
+    const context = this.canvas.getContext('2d');
+    this.paint(context);
+  }
+
+  componentDidUpdate() {
+    const context = this.canvas.getContext('2d');
+
+    // We only clear if the map is not shown
+    if (! this.state || ! this.state.image) {
+      context.clearRect(0, 0, MAP_SIZE, MAP_SIZE);
+    }
+
+    this.paint(context);
+  }
+
+  loadedImage(image) {
+    this.setState({ image });
+  }
+
+  mouseMoved(evt) {
+    console.log(evt.clientX, evt.clientY);
+  }
+
+  paint(context) {
+    context.save();
+    // Image data
+    if (this.state && this.state.image) {
+      const image = this.state.image;
+      context.globalCompositeOperation = 'source-over';
+      context.drawImage(image, 0, 0);
+    }
+
+    this.props.players.map((player) => {
+      context.fillStyle = '#' + player.id.slice(0, 6);
+      context.fillRect(player.x, player.y, 2, 2);
     });
+    context.restore();
   }
 
   render() {
-    return (
-      <div width='100%' style={styles.worldMap}>
-        { (this.state && this.state.hoveredPlayer) ?
-            <Stats
-              player={this.state.hoveredPlayer}
-              mapSize={2048}
-              /> :
-          null
-        }
-        <img
-          ref='worldmap'
-          src='CompanionWorldMap.png'
-          style={styles.worldMapImage}
-          alt='World Map'
-        />
-        {
-          this.props.players.map((player) => {
-            return (
-              <Player name={player.name}
-                      color={player.color}
-                      x={player.x}
-                      y={player.y}
-                      mapSize={2048}
-                      key={player.id}
-                      onMouseOver={this.playerHovered.bind(this, player)}
-              />
-            );
-          })
-        }
-      </div>
-    );
+    return <canvas
+            width={MAP_SIZE}
+            height={MAP_SIZE}
+            ref={(c) => this.canvas = c}
+            onMouseMove={this.mouseMoved}
+            />;
   }
 }
